refactor: migrate otel resource env to distro
