+++ distros/yamls/dotnet-legacy.yaml
@@ -13,5 +13,7 @@ spec:
     This distribution is for Dotnet applications running on old runtimes, using OpenTelemetry Native SDK and instrumentation libraries from the OpenTelemetry community.
   tiers: 
     - onprem
-  agentDirectories:
-    - directoryName: "{{ODIGOS_AGENTS_DIR}}/legacy-dotnet"
+  runtimeAgent:
+    directoryNames:
+      - "{{ODIGOS_AGENTS_DIR}}/legacy-dotnet"
+    k8sAttrsViaEnvVars: true
