+++ distros/yamls/python-enterprise.yaml
@@ -17,6 +17,7 @@ spec:
     - envName: PYTHONPATH
       envValue: '{{ODIGOS_AGENTS_DIR}}/python-ebpf:{{ODIGOS_AGENTS_DIR}}/python/opentelemetry/instrumentation/auto_instrumentation:{{ODIGOS_AGENTS_DIR}}/python'
       delimiter: ':'
-  agentDirectories:
-    - directoryName: "{{ODIGOS_AGENTS_DIR}}/python-ebpf"
-    - directoryName: "{{ODIGOS_AGENTS_DIR}}/python"
+  runtimeAgent:
+    directoryNames:
+      - "{{ODIGOS_AGENTS_DIR}}/python"
+      - "{{ODIGOS_AGENTS_DIR}}/python-ebpf"
